applied correct r_word_score
reservation word score uses "halstead difficulty" which is represented as a float value. Appropriate changes to reflect the usage of a float value has been made

+++ backend/output/models.py
@@ -29,7 +29,7 @@ class EfficiencyResult(models.Model):
     result = models.OneToOneField('output.Result', on_delete=models.PROTECT, related_name='efficiency_result')
     loc_score = models.IntegerField()
     control_flow_complexity_score = models.IntegerField()
-    reservation_words_score = models.IntegerField()
+    reservation_words_score = models.FloatField()
     data_flow_complexity_score = models.FloatField()
